LPS-55585 Rename package so that it doesn't conflict with the portal-impl one
