+++ modules/portal-security/portal-security-auth/src/com/liferay/portal/security/auth/impl/verifier/TunnelingServletAuthVerifier.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.auth.verifier;
+package com.liferay.portal.security.auth.impl.verifier;
 
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.exception.SystemException;
